requirements.txt
isort
lxml
jupyter
jira
pyyaml
matplotlib
pandas
gitpython
black
isort
matplotlib-venn
